Rework path printer
Previous incarnation of path printer used a bit of a shoddy and naive
recursive algorithm to make things work, and was greatly inefficient.
Start reworking the algorithm to make it more efficient via two main
changes:
* Implement a two-pass approach, first identifying cycles in the graph,
  then computing the paths and entropies
* Only recurse when finding a split

This commit introduces a first draft of the first main change.

Hopefully this make the algorithm workable for larger workloads.