Add `entropy_eval`
Add a separate utility to exhaustively compute the uncertainty
coefficient for LLVM snippets. Supports integer and float arithmetic
operations, as well as integer and float comparison operations (NB float
support is only for 16 bits for now).

Additionally add `plot.jl`, a Julia script to consume the CSV output of
the above binary, and produce plots to visually represent the data.

* Fix using base `clang++` to compile `libllvm_snippets.so`, instead
  prefering to use the same compiler as the project, via
  `CMAKE_CXX_COMPILER`. This helps when we have an old local `clang`,
  but a more-up-to-date one via `find_package`